requirements.txt
pandas>=2.2.0
numpy>=1.26.4
scikit-learn>=1.5.0
scipy>=1.12.0
scikit-surprise
streamlit>=1.32.0
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.20.0
nltk>=3.8.1
tqdm>=4.66.4